simplified hkl calc tests
